Feature/IterativeClosestPointFilter (#70)
* Added IterativeClosestPoint filter
* Added corresponding unit test
* Added util/nanoflann.hpp from DREAM3DReview
* Fixed unit test errors for InterpolatePointCloudToRegularGrid filter
* Fixed unit test errors for MapPointCloudToRegularGrid filter
* Fixed fmt::format build errors after updating vcpkg

+++ src/complex/Filter/Actions/CreateNeighborListAction.cpp
@@ -56,7 +56,7 @@ Result<> CreateNeighborListAction::apply(DataStructure& dataStructure, Mode mode
     return CreateNeighbors<bool>(dataStructure, m_TupleCount, m_Path, mode);
   }
   default:
-    throw std::runtime_error(fmt::format("CreateNeighborListAction: Invalid Numeric Type '{}'", m_Type));
+    throw std::runtime_error(fmt::format("CreateNeighborListAction: Invalid Numeric Type '{}'", static_cast<int8>(m_Type)));
   }
 }
